Threads causing excessive database use
If an application has many different clock tasks, and they are scheduled to run on a fairly frequent basis (every minute or so), the database connection is lost very quickly, if threading is enabled globally (via `config[:thread] = true` or otherwise).

After some investigation, I found that, when starting a new thread, it creates a new database connection.  If you do not close the connection to ActiveRecord, you will have a zombie process on the database.